test(Runtime Integration Tests): Outsourced mosaic test functions into its own file.
